+++ internal/providercache/installer.go
@@ -23,6 +23,7 @@ import (
 	"github.com/opentofu/opentofu/internal/depsfile"
 	"github.com/opentofu/opentofu/internal/getproviders"
 	"github.com/opentofu/opentofu/internal/tracing"
+	"github.com/opentofu/opentofu/internal/tracing/traceattrs"
 )
 
 // Installer is the main type in this package, representing a provider installer
@@ -445,8 +446,8 @@ func (i *Installer) ensureProviderVersionsInstall(
 		traceCtx, span := tracing.Tracer().Start(ctx,
 			"Install Provider",
 			trace.WithAttributes(
-				otelAttr.String(tracing.ProviderAddressAttributeName, provider.String()),
-				otelAttr.String(tracing.ProviderVersionAttributeName, version.String()),
+				otelAttr.String(traceattrs.ProviderAddress, provider.String()),
+				otelAttr.String(traceattrs.ProviderVersion, version.String()),
 			),
 		)
 
@@ -728,15 +729,15 @@ func tryInstallPackageFromCacheDir(
 	preferredHashes []getproviders.Hash,
 	mayBreakDependencyLockFile bool,
 
-// FIXME: The above set of arguments came from exhaustively including
-// everything that a previously-inline version of this chunk of code
-// referred to, to minimize the risk of factoring it out. In future
-// we should try to separate these concerns a little better so that
-// this doesn't need so many arguments. For example, it might be better
-// for the caller to be responsible for updating "locks" when
-// installation is successful, but that would likely require changing
-// the order of emitted events so that the locks-update event
-// comes after the successful-linking event.
+	// FIXME: The above set of arguments came from exhaustively including
+	// everything that a previously-inline version of this chunk of code
+	// referred to, to minimize the risk of factoring it out. In future
+	// we should try to separate these concerns a little better so that
+	// this doesn't need so many arguments. For example, it might be better
+	// for the caller to be responsible for updating "locks" when
+	// installation is successful, but that would likely require changing
+	// the order of emitted events so that the locks-update event
+	// comes after the successful-linking event.
 ) (installed bool, err error) {
 	evts := installerEventsForContext(ctx)
