+++ internal/tracing/traceattrs/traceattrs.go
@@ -3,11 +3,11 @@
 // Copyright (c) 2023 HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package tracing
+package traceattrs
 
 const (
 	// Common attributes names used across the codebase
 
-	ProviderAddressAttributeName = "opentofu.provider.address"
-	ProviderVersionAttributeName = "opentofu.provider.version"
+	ProviderAddress = "opentofu.provider.address"
+	ProviderVersion = "opentofu.provider.version"
 )
